Create a new popup type
