+++ src/app/_components/_sharedcomponents/Popup/Popup.tsx
@@ -2,8 +2,9 @@
 import { PopupData, PopupType, usePopup } from '@/app/_contexts/Popup.context';
 import { Box, SxProps, Theme } from '@mui/material';
 import React from 'react';
-import { DefaultPopup, DropdownPopup, NumberPopup, PilePopup, SelectCardsPopup } from './Popup.types';
+import { ActionTriggerPopup, DefaultPopup, DropdownPopup, NumberPopup, PilePopup, SelectCardsPopup } from './Popup.types';
 import { DefaultPopupModal } from './PopupVariant/DefaultPopup';
+import { ActionTriggerPopupModal } from './PopupVariant/ActionTriggersPopup';
 import { PilePopupModal } from './PopupVariant/PilePopup';
 import { SelectCardsPopupModal } from './PopupVariant/SelectCardsPopup';
 import { contentStyle } from './Popup.styles';
@@ -64,13 +65,15 @@ const PopupShell: React.FC<IPopupShellProps> = ({
     sidebarOpen = false
 }) => {
     const { popups, focusPopup } = usePopup();
-    const { connectedPlayer }= useGame();
+    const { connectedPlayer } = useGame();
     const isPilePopup = (popup: PopupData): popup is PilePopup => popup.type === 'pile';
 
     if (popups.length === 0) return null; // No popup to display
 
     const renderPopupContent = (type: PopupType, data: PopupData) => {
         switch (type) {
+            case 'actionTrigger':
+                return <ActionTriggerPopupModal data={data as ActionTriggerPopup} />;
             case 'default':
                 return <DefaultPopupModal data={data as DefaultPopup} />;
             case 'pile':
@@ -101,8 +104,8 @@ const PopupShell: React.FC<IPopupShellProps> = ({
     }
 
     const [nonDefaultPopups, defaultPopups] = [
-        popups.filter((popup) => popup.type !== 'default'),
-        popups.filter((popup) => popup.type === 'default')
+        popups.filter((popup) => popup.type !== 'default' && popup.type !== 'actionTrigger'),
+        popups.filter((popup) => popup.type === 'default' || popup.type === 'actionTrigger')
     ];
 
     const overlayStyle = {
