+++ src/app/_components/_sharedcomponents/Popup/PopupVariant/DefaultPopup.tsx
@@ -1,6 +1,5 @@
 import { useGame } from '@/app/_contexts/Game.context';
-import VisibilityIcon from '@mui/icons-material/Visibility';
-import { Box, IconButton, Popover, Typography } from '@mui/material';
+import { Box, IconButton, Typography } from '@mui/material';
 import { MouseEvent, useState } from 'react';
 import { BiMinus, BiPlus } from 'react-icons/bi';
 import GradientBorderButton from '@/app/_components/_sharedcomponents/_styledcomponents/GradientBorderButton';
@@ -14,55 +13,14 @@ import {
 } from '../Popup.styles';
 import { DefaultPopup, PopupButton } from '../Popup.types';
 import RichText from '../../RichText/RichText';
-import { CardStyle } from '../../Cards/CardTypes';
-import { s3CardImageURL } from '@/app/_utils/s3Utils';
-import { useCardImageLocale } from '@/app/_contexts/CardImageLocale.context';
 
 interface ButtonProps {
     data: DefaultPopup;
 }
 
 export const DefaultPopupModal = ({ data }: ButtonProps) => {
-    const { sendGameMessage, gameState, connectedPlayer } = useGame();
-    const locale = useCardImageLocale();
+    const { sendGameMessage } = useGame();
     const [isMinimized, setIsMinimized] = useState(false);
-    const [previewAnchorElement, setPreviewAnchorElement] = useState<HTMLElement | null>(null);
-    const [previewImage, setPreviewImage] = useState<string | null>(null);
-
-    const triggerWindow = gameState?.players[connectedPlayer]?.promptState?.promptType === 'triggerWindow';
-    const previewOpen = Boolean(previewAnchorElement);
-
-    const getButtonBackgroundImage = (button: PopupButton): string | undefined => {
-        if (!button.sourceCardSetId) {
-            return undefined;
-        }
-
-        return s3CardImageURL(
-            { setId: button.sourceCardSetId, type: '', id: '' },
-            locale,
-            CardStyle.Plain
-        );
-    };
-
-    const handlePreviewOpen = (event: MouseEvent<HTMLElement>, button: PopupButton) => {
-        if (!button.sourceCardSetId) {
-            return;
-        }
-
-        const imageUrl = s3CardImageURL(
-            { setId: button.sourceCardSetId, type: '', id: '' },
-            locale,
-            CardStyle.Plain
-        );
-
-        setPreviewAnchorElement(event.currentTarget);
-        setPreviewImage(`url(${imageUrl})`);
-    };
-
-    const handlePreviewClose = () => {
-        setPreviewAnchorElement(null);
-        setPreviewImage(null);
-    };
 
     const renderPopupContent = () => {
         if (isMinimized) return null;
@@ -71,118 +29,8 @@ export const DefaultPopupModal = ({ data }: ButtonProps) => {
                 {data.description && (
                     <RichText text={data.description} sx={textStyle} component={Typography}/>
                 )}
-                <Box sx={triggerWindow ? {
-                    ...footerStyle,
-                    flexDirection: 'row',
-                    justifyContent: 'center',
-                    alignItems: 'stretch',
-                    overflowX: 'auto',
-                    overflowY: 'hidden',
-                    maxWidth: 'min(80vw, 58rem)',
-                    paddingBottom: '0.5rem',
-                    marginTop: '1rem',
-                    marginBottom: '2rem',
-                    paddingRight: { xs: '2rem', md: 0 }
-                } : footerStyle}>
+                <Box sx={footerStyle}>
                     {data.buttons.map((button: PopupButton, index: number) => {
-                        const backgroundImage = triggerWindow ? getButtonBackgroundImage(button) : undefined;
-                        const triggerButtonSx = triggerWindow ? {
-                            width: '100%',
-                            height: '100%',
-                            padding: '0.75rem',
-                            alignItems: 'flex-end',
-                            justifyContent: 'center',
-                            textAlign: 'center',
-                            whiteSpace: 'normal',
-                            border: 0,
-                            borderRadius: '13px',
-                            background: backgroundImage
-                                ? `linear-gradient(to top, rgba(3, 12, 19, 0.96) 0%, rgba(3, 12, 19, 0.9) 45%, rgba(15, 31, 39, 0.58) 76%, rgba(15, 31, 39, 0.34) 100%), url(${backgroundImage})`
-                                : '#1E2D32',
-                            backgroundPosition: 'center top',
-                            backgroundSize: 'cover',
-                            fontSize: { xs: '11px', md: '15px' },
-                            '&:hover': {
-                                filter: 'brightness(1.15)',
-                            },
-                            '& .MuiTypography-root': {
-                                fontSize: '0.78rem',
-                                lineHeight: 1.2,
-                                textShadow: '0 1px 3px rgba(0, 0, 0, 0.95)',
-                            },
-                        } : undefined;
-
-                        const triggerButton = (
-                            <GradientBorderButton
-                                fillColor={button.selected ? 'rgba(102, 229, 255, 0.2)' : undefined}
-                                sx={triggerButtonSx}
-                                onClickHandler={() => {
-                                    sendGameMessage([button.command, button.arg, button.uuid]);
-                                }}
-                            >
-                                <RichText text={button.text} />
-                            </GradientBorderButton>
-                        );
-
-                        if (triggerWindow) {
-                            return (
-                                <Box
-                                    key={`${button.uuid}:${index}`}
-                                    sx={{
-                                        position: 'relative',
-                                        overflow: 'hidden',
-                                        aspectRatio: '1 / 1.4',
-                                        flex: '0 0 clamp(116px, 16vw, 10rem)',
-                                        width: 'clamp(116px, 16vw, 10rem)',
-                                        padding: '2px',
-                                        borderRadius: '15px',
-                                        border: '1px solid #21313ac2',
-                                        '@media (hover: hover) and (pointer: fine)': {
-                                            '& .trigger-card-preview-button': {
-                                                opacity: 0,
-                                                pointerEvents: 'none',
-                                            },
-                                            '&:hover .trigger-card-preview-button': {
-                                                opacity: 1,
-                                                pointerEvents: 'auto',
-                                            },
-                                        },
-                                    }}
-                                >
-                                    {triggerButton}
-                                    {button.sourceCardSetId && (
-                                        <IconButton
-                                            className="trigger-card-preview-button"
-                                            aria-label="View card details"
-                                            size="small"
-                                            onClick={(event) => {
-                                                event.stopPropagation();
-                                                handlePreviewOpen(event, button);
-                                            }}
-                                            onMouseEnter={(event) => handlePreviewOpen(event, button)}
-                                            onMouseLeave={handlePreviewClose}
-                                            sx={{
-                                                position: 'absolute',
-                                                top: '0.35rem',
-                                                right: '0.35rem',
-                                                zIndex: 2,
-                                                color: 'white',
-                                                backgroundColor: 'rgba(3, 12, 19, 0.72)',
-                                                border: '1px solid rgba(255, 255, 255, 0.38)',
-                                                boxShadow: '0 2px 6px rgba(0, 0, 0, 0.55)',
-                                                transition: 'opacity 140ms ease, background-color 140ms ease',
-                                                '&:hover': {
-                                                    backgroundColor: 'rgba(3, 12, 19, 0.9)',
-                                                },
-                                            }}
-                                        >
-                                            <VisibilityIcon fontSize="small" />
-                                        </IconButton>
-                                    )}
-                                </Box>
-                            );
-                        }
-
                         return (
                             <GradientBorderButton
                                 key={`${button.uuid}:${index}`}
@@ -196,34 +44,6 @@ export const DefaultPopupModal = ({ data }: ButtonProps) => {
                         );
                     })}
                 </Box>
-                <Popover
-                    id="trigger-card-preview-popover"
-                    sx={{ pointerEvents: 'none' }}
-                    open={previewOpen}
-                    anchorEl={previewAnchorElement}
-                    onClose={handlePreviewClose}
-                    disableRestoreFocus
-                    slotProps={{ paper: { sx: { backgroundColor: 'transparent', boxShadow: 'none' }, tabIndex: -1 } }}
-                    anchorOrigin={{
-                        vertical: 'center',
-                        horizontal: 'right',
-                    }}
-                    transformOrigin={{
-                        vertical: 'center',
-                        horizontal: 'left',
-                    }}
-                >
-                    <Box
-                        sx={{
-                            borderRadius: '.38em',
-                            backgroundImage: previewImage,
-                            backgroundSize: 'cover',
-                            backgroundRepeat: 'no-repeat',
-                            aspectRatio: '1 / 1.4',
-                            width: 'clamp(200px, 60vw, 16rem)',
-                        }}
-                    />
-                </Popover>
             </>
         );
     };
